% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LinearRegression.R
\name{addCovariateSetBySet}
\alias{addCovariateSetBySet}
\title{A function to Add variable set after set}
\usage{
addCovariateSetBySet(outcome, target, covariate)

addCovariateSetBySet(outcome, target, covariate)
}
\arguments{
\item{outcome}{Linear model set outcome. It could be a character or a character vector}

\item{target}{variable of interest}

\item{covariate}{control variables in your model. It should be a list of vectors}
}
\value{
regression table with reference level added

regression table with reference level added
}
\description{
This function allows you to add variable set into your model set by set

This function allows you to add variable set into your model set by set
}
\examples{
model <- lm(formula = Sepal.Width ~ Petal.Length, data = iris)
clean.lm(model)
outcome <- "outcome"
target <- "V1"
covariate <- list(c("V2", "V3"), c("V5", "V6"))
addCovariateSetBySet(outcome = outcome, target = target, covariate = covariate)
}
\keyword{Model}
