% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/broom218-package.R
\docType{package}
\name{broom218-package}
\alias{broom218}
\alias{broom218-package}
\title{broom218: clean result from tidy function in package broom}
\description{
This package is used to clean data from tidy function in broom package to make sure the regression table meets requirements of statistic table Use four spaces when indenting paragraphs within the Description.
}
\keyword{internal}
