% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LinearRegression.R
\name{DetermineModelType}
\alias{DetermineModelType}
\title{A function determine model type}
\usage{
DetermineModelType(model)

DetermineModelType(model)
}
\arguments{
\item{model}{The model passed to the function}
}
\value{
The type of the object passed into this function. Defaults to Linear Regression and Logistic Regression

The type of the object passed into this function. Defaults to Linear Regression and Logistic Regression
}
\description{
This function allows you to find what kind a model is.

This function allows you to find what kind a model is.
}
\examples{
model <- lm(formula = Sepal.Width ~ Petal.Length, data = iris)
DetermineModelType(model)
model <- lm(formula = Sepal.Width ~ Petal.Length, data = iris)
DetermineModelType(model)
}
\keyword{Model}
