% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LinearRegression.R
\name{regressionset.lm}
\alias{regressionset.lm}
\title{A function to summary linear model regression set table}
\usage{
regressionset.lm(
  data,
  outcome,
  target,
  covariate,
  pType = "withPValue",
  estimateType = "betase",
  digits = 2,
  pDigits = 4,
  strata = NULL,
  ...
)

regressionset.lm(
  data,
  outcome,
  target,
  covariate,
  pType = "withPValue",
  estimateType = "betase",
  digits = 2,
  pDigits = 4,
  strata = NULL,
  ...
)
}
\arguments{
\item{outcome}{Linear model set outcome. It could be a character vector or a list}

\item{target}{variable of interest. For now the code only supports one target variable of interest}

\item{covariate}{control variables in your model. It should be a list of vectors}

\item{pType}{whether put P into final table, default to withPValue, another selection is withMark}

\item{estimateType}{Linear regression table type, defaults to "betase"}

\item{digits}{Digit number of the estimate}

\item{pDigits}{Digit number of the P value}

\item{...}{other parameters to pass into tidy function from broom}
}
\value{
regression table with reference level added

regression table with reference level added
}
\description{
This function allows you to generate linear regression table

This function allows you to generate linear regression table
}
\examples{
model <- lm(formula = Sepal.Width ~ Petal.Length, data = iris)
clean.lm(model)
model <- lm(formula = Sepal.Width ~ Petal.Length, data = iris)
clean.lm(model)
}
\keyword{Model}
